+++ app/Http/Controllers/Backend/ExamController.php
@@ -6,8 +6,8 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\Exam;
 use Illuminate\Support\Facades\Storage;
-use App\Http\Requests\ExamPostRequest;
-use App\Http\Requests\ExamPutRequest;
+use App\Http\Requests\Backend\ExamPostRequest;
+use App\Http\Requests\Backend\ExamPutRequest;
 use Session;
 
 class ExamController extends Controller
