+++ app/Http/Requests/Backend/CategoryRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Backend;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class ExamPostRequest extends FormRequest
+class CategoryRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,8 +24,7 @@ class ExamPostRequest extends FormRequest
     public function rules()
     {
         return [
-            'title' => 'required',
-            'audio' => 'required|mimes:mp3,mpga'
+            'name' => 'required',
         ];
     }
 }
